@@ -1,4 +1,4 @@
-package com.tutoringplatform.config;
+package com.tutoringplatform.security;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -20,9 +20,9 @@ public class SecurityConfig {
     @Bean
     public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
         http
-                .csrf().disable()
+                .csrf(csrf -> csrf.disable())
                 .authorizeHttpRequests(auth -> auth
-                        .antMatchers("/**").permitAll() 
+                        .requestMatchers("/**").permitAll() 
                         .anyRequest().authenticated()
                 );
         return http.build();
