@@ -1,6 +1,6 @@
-package com.tutoringplatform.repositories.interfaces;
+package com.tutoringplatform.file;
 
-import com.tutoringplatform.models.FileMetaData;
+import com.tutoringplatform.shared.IRepository;
 import java.util.List;
 
 public interface IFileRepository extends IRepository<FileMetaData> {
