@@ -1,6 +1,6 @@
-package com.tutoringplatform.repositories.interfaces;
+package com.tutoringplatform.payment;
 
-import com.tutoringplatform.models.Payment;
+import com.tutoringplatform.shared.IRepository;
 
 public interface IPaymentRepository extends IRepository<Payment> {
     Payment findByBookingId(String bookingId);
