@@ -1,10 +1,10 @@
-package com.tutoringplatform.dto.response;
+package com.tutoringplatform.shared.dto.response;
 
 import java.time.LocalDateTime;
 import java.util.List;
 
-import com.tutoringplatform.models.availability.AvailabilityException;
-import com.tutoringplatform.models.availability.RecurringAvailability;
+import com.tutoringplatform.user.availability.model.AvailabilityException;
+import com.tutoringplatform.user.availability.model.RecurringAvailability;
 
 public class AvailabilityResponse {
     private String tutorId;
