@@ -1,6 +1,6 @@
-package com.tutoringplatform.repositories.interfaces;
+package com.tutoringplatform.user;
 
-import com.tutoringplatform.models.User;
+import com.tutoringplatform.shared.IRepository;
 
 public interface IUserRepository<T extends User> extends IRepository<T> {
     T findByEmail(String email);
