@@ -1,8 +1,7 @@
-package com.tutoringplatform.controllers;
+package com.tutoringplatform.user.dashboard;
 
-import com.tutoringplatform.dto.response.StudentDashboardResponse;
-import com.tutoringplatform.dto.response.TutorDashboardResponse;
-import com.tutoringplatform.services.DashboardService;
+import com.tutoringplatform.shared.dto.response.StudentDashboardResponse;
+import com.tutoringplatform.shared.dto.response.TutorDashboardResponse;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.http.ResponseEntity;
