@@ -1,4 +1,4 @@
-package com.tutoringplatform.repositories.impl;
+package com.tutoringplatform.user.student;
 
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -7,9 +7,6 @@ import java.util.Map;
 import java.util.stream.Collectors;
 
 import org.springframework.stereotype.Repository;
-
-import com.tutoringplatform.models.Student;
-import com.tutoringplatform.repositories.interfaces.IStudentRepository;
 
 @Repository
 public class StudentRepository implements IStudentRepository {
